Fix some warnings.

+++ asio/include/asio/detail/win_iocp_socket_service.hpp
@@ -1261,7 +1261,7 @@ public:
     // Receive some data.
     DWORD bytes_transferred = 0;
     DWORD recv_flags = flags;
-    int endpoint_size = sender_endpoint.capacity();
+    int endpoint_size = static_cast<int>(sender_endpoint.capacity());
     int result = ::WSARecvFrom(impl.socket_, bufs, i, &bytes_transferred,
         &recv_flags, sender_endpoint.data(), &endpoint_size, 0, 0);
     if (result != 0)
@@ -1279,7 +1279,7 @@ public:
       return 0;
     }
 
-    sender_endpoint.resize(endpoint_size);
+    sender_endpoint.resize(static_cast<std::size_t>(endpoint_size));
 
     ec = asio::error_code();
     return bytes_transferred;
@@ -1299,7 +1299,7 @@ public:
           &receive_from_operation<
             MutableBufferSequence, Handler>::destroy_impl),
         endpoint_(endpoint),
-        endpoint_size_(endpoint.capacity()),
+        endpoint_size_(static_cast<int>(endpoint.capacity())),
         work_(io_service),
         buffers_(buffers),
         handler_(handler)
@@ -1618,7 +1618,8 @@ public:
         GetAcceptExSockaddrs(handler_op->output_buffer(), 0,
             handler_op->address_length(), handler_op->address_length(),
             &local_addr, &local_addr_length, &remote_addr, &remote_addr_length);
-        if (remote_addr_length > peer_endpoint.capacity())
+        if (static_cast<std::size_t>(remote_addr_length)
+            > peer_endpoint.capacity())
         {
           last_error = WSAEINVAL;
         }
@@ -1626,7 +1627,7 @@ public:
         {
           using namespace std; // For memcpy.
           memcpy(peer_endpoint.data(), remote_addr, remote_addr_length);
-          peer_endpoint.resize(remote_addr_length);
+          peer_endpoint.resize(static_cast<std::size_t>(remote_addr_length));
         }
       }
